Project grid on css modules
