+++ src/components/ProjectsGrid/Filter.module.css
@@ -1,0 +1,15 @@
+.wrapper {
+  display: flex;
+  justify-content: space-evenly;
+  padding-bottom: 27px;
+}
+
+.link {
+  color: #000;
+  font-size: 35px;
+
+  &:hover,
+  &.active {
+    text-decoration: line-through;
+  }
+}
