+++ src/components/index.ts
@@ -5,7 +5,7 @@ export { Header } from "./Header";
 export { Layout } from "./Layout";
 export { Paragraph } from "./Paragraph";
 export { ProjectDetail } from "./ProjectDetail";
-export { ProjectsGrid } from "./ProjectsGrid";
+export * from "./ProjectsGrid";
 export { ProjectsList } from "./ProjectsList";
 export { RootRedirect } from "./RootRedirect";
 export { WrappedAsset } from "./WrappedAsset";
